abdelkareem

+++ week3/Admin/blogs/delete.php
@@ -2,6 +2,7 @@
 ########################################################################################################
 require '../helpers/db.php';
 require '../helpers/functions.php';
+require '../helpers/checklogin.php';
 
 # Fetch Id .... 
 $id = $_GET['id']; 
@@ -11,10 +12,24 @@ $id = $_GET['id'];
 if(!validate($id,'int')){
     $message = ["Error" => "Invalid Id"];
 }else{
+    
     # Fetch image name .... 
     $sql  = "select image from blogs where id = $id"; 
     $op   = doQuery($sql); 
-    $data =  mysqli_fetch_assoc($op); 
+    $data =  mysqli_fetch_assoc($op);
+
+
+
+
+    if (!(($_SESSION['user']['role_id'] == 2) || ($_SESSION['user']['role_id'] == 5 && $_SESSION['user']['id'] == $data['addedBy']))) {
+        $message = ["Error" => "Can't complete this Operation"];
+
+        # Set Session ... 
+        $_SESSION['Message'] = $message;
+
+        header("Location: ".url('/blogs/'));
+        exit();
+    }
 
 
     $sql = "delete from blogs where id = $id";
